Move DeviceApiService implementation directly in JsonApiRequestHandler
Also get rid of network related Crashlytics calls.
